+++ termenu/__init__.py
@@ -1,0 +1,1 @@
+from .termenu import *
